added seller board and add new book successfully
